requirements.txt
beautifulsoup4
pyyaml
requests
unidecode